Fixed the login logic so that either the logged in username is visible, or the signed in user is listed as anonymouns. Refactored all the css files into a css directory. Deleted the Users component as it was not being used.
