+++ client/src/AddPrep.js
@@ -1,5 +1,5 @@
 import React, { Component } from "react"
-import "./AddPrep.css"
+import "./css/AddPrep.css"
 
 class AddPrep extends Component {
