+++ client/src/Keywords.js
@@ -1,5 +1,5 @@
 import React, { Component } from "react"
-import "./Keywords.css"
+import "./css/Keywords.css"
 
 class Keywords extends Component {
