+++ client/src/Landing.js
@@ -1,6 +1,6 @@
 import React from 'react'
 import NavBar from "./NavBar"
-import "./Landing.css"
+import "./css/Landing.css"
 import { Link } from "react-router-dom"
 
 function Landing() {
