--- client/src/Users.js
@@ -1,9 +1,0 @@
-import React from 'react'
-
-export default function Users() {
-   return (
-      <div>
-         <h1>User Component is up Man!</h1>
-      </div>
-   )
-}
