+++ client/src/index.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import { BrowserRouter as Router } from "react-router-dom"
-import './index.css';
+import './css/index.css';
 import App from './App';
 
 ReactDOM.render(
